backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
daphne==4.1.2
python-dotenv==1.0.1
celery==5.4.0
pycryptodome==3.20.0
Pillow==10.4.0
redis==5.0.4
channels-redis==4.2.0
